+++ src/VirtoCommerce.ExperienceApiModule.DigitalCatalog/Queries/LoadRelatedCatalogOutlineQueryHandler.cs
@@ -26,28 +26,4 @@ namespace VirtoCommerce.XDigitalCatalog.Queries
             };
         }
     }
-
-    public class LoadRelatedSlugPathQueryHandler : IQueryHandler<LoadRelatedSlugPathQuery, LoadRelatedSlugPathResponse>
-    {
-        private readonly IStoreService _storeService;
-
-        public LoadRelatedSlugPathQueryHandler(IStoreService storeService)
-        {
-            _storeService = storeService;
-        }
-
-        public async Task<LoadRelatedSlugPathResponse> Handle(LoadRelatedSlugPathQuery request, CancellationToken cancellationToken)
-        {
-            var store = await _storeService.GetByIdAsync(request.StoreId);
-            if (store is null) return null;
-
-            var language = request.CultureName ?? store.DefaultLanguage;
-            var slug = request.Outlines.GetSeoPath(store, language, null);
-
-            return new LoadRelatedSlugPathResponse
-            {
-                Slug = slug
-            };
-        }
-    }
 }
